+++ test/basic_test.dart
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import 'package:dartastic_opentelemetry/dartastic_opentelemetry.dart';
+import 'package:dartastic_opentelemetry/src/version.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -14,6 +15,25 @@ void main() {
     final attrs = OTel.defaultResource!.attributes.toList();
     final serviceName = attrs.firstWhere((a) => a.key == 'service.name');
     expect(serviceName.value, equals('test-service'));
+
+    await OTel.reset();
+  });
+
+  test('Default resource includes telemetry.sdk.* attributes', () async {
+    await OTel.initialize(serviceName: 'test-service');
+
+    final attrs = OTel.defaultResource!.attributes.toList();
+
+    final sdkLanguage =
+        attrs.firstWhere((a) => a.key == 'telemetry.sdk.language');
+    expect(sdkLanguage.value, equals('dart'));
+
+    final sdkName = attrs.firstWhere((a) => a.key == 'telemetry.sdk.name');
+    expect(sdkName.value, equals('opentelemetry'));
+
+    final sdkVersion =
+        attrs.firstWhere((a) => a.key == 'telemetry.sdk.version');
+    expect(sdkVersion.value, equals(packageVersion));
 
     await OTel.reset();
   });